Re-config navigator for better management
