+++ mobile/src/navigation/AuthNavigator.tsx
@@ -1,0 +1,15 @@
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { AuthStackParamList } from "../types/navigation";
+import LoginScreen from "../screens/Auth/LoginScreen";
+import RegisterScreen from "../screens/Auth/RegisterScreen";
+
+const Stack = createNativeStackNavigator<AuthStackParamList>();
+
+const AuthNavigator = () => (
+  <Stack.Navigator>
+    <Stack.Screen name="Login" component={LoginScreen} />
+    <Stack.Screen name="Register" component={RegisterScreen} />
+  </Stack.Navigator>
+);
+
+export default AuthNavigator;
